make proper use of package
