+++ squeakernet/hx711.py
