+++ squeakernet/logcategory.py
@@ -1,0 +1,4 @@
+class LogCategory(Enum):
+    FEED = 1
+    SYSTEM = 2
+    WEIGHT = 3
