+++ squeakernet/reading.py
