+++ squeakernet/scale.py
@@ -2,6 +2,8 @@ import os
 import sys
 import ConfigParser
 from hx711 import HX711
+import db
+from logcategory import LogCategory
 
 config = ConfigParser.ConfigParser()
 config.read(os.path.join(sys.path[0], "squeakernet.ini"))
@@ -26,3 +28,20 @@ def get_weight():
         hx.cleanup()
 
     return result
+
+def log_weight(value = None):
+    if value is None:
+        value =  get_weight()
+    if value is None:
+        print 'No log: Bad reading from scale.'
+        return
+
+    last_weight = db.get_last_log(LogCategory.WEIGHT)
+    change = value - last_weight.reading
+    log_threshold = 1
+
+    if(abs(change)) > log_threshold:
+        db.write_log(LogCategory.WEIGHT, '', value)
+        print 'Logged weight of %s to the database (change of %s).' % (value, change)
+    else:
+        print 'No log: weight change of %s is < %s' % (change, log_threshold)
