+++ squeakernet/speech.py
